azure/requirement.txt
azure-cognitiveservices-speech = 1.47.0
python-dotenv = 1.2.1
